api/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-jose==3.3.0
python-multipart==0.0.20